fix review items
